requirements.txt
discord.py
aiowebostv
loguru
pyyaml